+++ firebase-dataconnect/src/test/kotlin/com/google/firebase/dataconnect/sqlite/DataConnectCacheDatabaseUnitTest.kt
@@ -19,12 +19,16 @@ package com.google.firebase.dataconnect.sqlite
 import android.database.sqlite.SQLiteDatabase
 import com.google.firebase.dataconnect.core.DataConnectAuth.AuthUid
 import com.google.firebase.dataconnect.core.Logger
+import com.google.firebase.dataconnect.core.QueryId
 import com.google.firebase.dataconnect.sqlite.DataConnectCacheDatabase.GetQueryResultResult
 import com.google.firebase.dataconnect.sqlite.DataConnectCacheDatabase.GetQueryResultResult.Found
 import com.google.firebase.dataconnect.sqlite.DataConnectCacheDatabase.GetQueryResultResult.NotFound
 import com.google.firebase.dataconnect.sqlite.DataConnectCacheDatabase.GetQueryResultResult.Stale
+import com.google.firebase.dataconnect.sqlite.DataConnectCacheDatabase.SqliteSequenceNumber
 import com.google.firebase.dataconnect.sqlite.QueryResultArb.EntityRepeatPolicy.INTER_SAMPLE
 import com.google.firebase.dataconnect.sqlite.QueryResultArb.EntityRepeatPolicy.INTER_SAMPLE_MUTATED
+import com.google.firebase.dataconnect.sqlite.SQLiteDatabaseExts.execSQL
+import com.google.firebase.dataconnect.sqlite.SQLiteDatabaseExts.rawQuery
 import com.google.firebase.dataconnect.testutil.CleanupsRule
 import com.google.firebase.dataconnect.testutil.DataConnectLogLevelRule
 import com.google.firebase.dataconnect.testutil.Quadruple
@@ -52,23 +56,27 @@ import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.assertions.withClue
 import io.kotest.common.DelicateKotest
 import io.kotest.common.ExperimentalKotest
+import io.kotest.matchers.nulls.shouldBeNull
 import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.types.shouldBeInstanceOf
 import io.kotest.property.Arb
 import io.kotest.property.EdgeConfig
 import io.kotest.property.PropTestConfig
+import io.kotest.property.PropertyContext
 import io.kotest.property.ShrinkingMode
 import io.kotest.property.arbitrary.bind
 import io.kotest.property.arbitrary.byte
 import io.kotest.property.arbitrary.byteArray
 import io.kotest.property.arbitrary.distinct
+import io.kotest.property.arbitrary.filter
 import io.kotest.property.arbitrary.filterNot
 import io.kotest.property.arbitrary.flatMap
 import io.kotest.property.arbitrary.int
 import io.kotest.property.arbitrary.map
 import io.kotest.property.arbitrary.of
 import io.kotest.property.arbitrary.orNull
+import io.kotest.property.assume
 import io.kotest.property.checkAll
 import io.mockk.CapturingSlot
 import io.mockk.every
@@ -83,6 +91,7 @@ import kotlin.reflect.KClass
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.milliseconds
 import kotlin.time.Duration.Companion.nanoseconds
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -110,7 +119,7 @@ class DataConnectCacheDatabaseUnitTest {
   private val lazyDataConnectCacheDatabase = lazy {
     val dbFile = File(temporaryFolder.newFolder(), "db.sqlite")
     val mockLogger: Logger = mockk(relaxed = true)
-    DataConnectCacheDatabase(dbFile, mockLogger)
+    DataConnectCacheDatabase(dbFile, Dispatchers.Default, mockLogger)
   }
 
   private val dataConnectCacheDatabase: DataConnectCacheDatabase by
@@ -132,7 +141,7 @@ class DataConnectCacheDatabaseUnitTest {
   fun `initialize() should create the database at the file given to the constructor`() = runTest {
     val dbFile = File(temporaryFolder.newFolder(), "db.sqlite")
     val mockLogger: Logger = mockk(relaxed = true)
-    val dataConnectCacheDatabase = DataConnectCacheDatabase(dbFile, mockLogger)
+    val dataConnectCacheDatabase = DataConnectCacheDatabase(dbFile, Dispatchers.Default, mockLogger)
 
     dataConnectCacheDatabase.initialize()
 
@@ -150,7 +159,7 @@ class DataConnectCacheDatabaseUnitTest {
   fun `initialize() should create an in-memory database if a null file given to the constructor`() =
     runTest {
       val mockLogger: Logger = mockk(relaxed = true)
-      val dataConnectCacheDatabase = DataConnectCacheDatabase(null, mockLogger)
+      val dataConnectCacheDatabase = DataConnectCacheDatabase(null, Dispatchers.Default, mockLogger)
 
       dataConnectCacheDatabase.initialize()
 
@@ -278,7 +287,7 @@ class DataConnectCacheDatabaseUnitTest {
     ) { authUid, queryId, maxAge, structSample ->
       dataConnectCacheDatabase.insertQueryResult(
         authUid.authUid,
-        queryId.bytes,
+        queryId.queryId,
         structSample.struct,
         maxAge = maxAge,
         currentTimeMillis = 0L,
@@ -288,7 +297,7 @@ class DataConnectCacheDatabaseUnitTest {
       val result =
         dataConnectCacheDatabase.getQueryResult(
           authUid.authUid,
-          queryId.bytes,
+          queryId.queryId,
           currentTimeMillis = 0L,
           staleResult = Stale::class,
         )
@@ -311,7 +320,7 @@ class DataConnectCacheDatabaseUnitTest {
       structSamples.forEach {
         dataConnectCacheDatabase.insertQueryResult(
           authUid.authUid,
-          queryId.bytes,
+          queryId.queryId,
           it.struct,
           maxAge = maxAge,
           currentTimeMillis = 0L,
@@ -322,7 +331,7 @@ class DataConnectCacheDatabaseUnitTest {
       val result =
         dataConnectCacheDatabase.getQueryResult(
           authUid.authUid,
-          queryId.bytes,
+          queryId.queryId,
           currentTimeMillis = 0L,
           staleResult = Stale::class,
         )
@@ -344,7 +353,7 @@ class DataConnectCacheDatabaseUnitTest {
     ) { (authUid1, authUid2), queryId, maxAge, (structSample1, structSample2) ->
       dataConnectCacheDatabase.insertQueryResult(
         authUid1.authUid,
-        queryId.bytes,
+        queryId.queryId,
         structSample1.struct,
         maxAge = maxAge,
         currentTimeMillis = 0L,
@@ -352,7 +361,7 @@ class DataConnectCacheDatabaseUnitTest {
       )
       dataConnectCacheDatabase.insertQueryResult(
         authUid2.authUid,
-        queryId.bytes,
+        queryId.queryId,
         structSample2.struct,
         maxAge = maxAge,
         currentTimeMillis = 0L,
@@ -362,7 +371,7 @@ class DataConnectCacheDatabaseUnitTest {
       val result1 =
         dataConnectCacheDatabase.getQueryResult(
           authUid1.authUid,
-          queryId.bytes,
+          queryId.queryId,
           currentTimeMillis = 0L,
           staleResult = Stale::class,
         )
@@ -370,7 +379,7 @@ class DataConnectCacheDatabaseUnitTest {
       val result2 =
         dataConnectCacheDatabase.getQueryResult(
           authUid2.authUid,
-          queryId.bytes,
+          queryId.queryId,
           currentTimeMillis = 0L,
           staleResult = Stale::class,
         )
@@ -395,7 +404,7 @@ class DataConnectCacheDatabaseUnitTest {
     ) { authUid, (queryId1, queryId2), maxAge, (structSample1, structSample2) ->
       dataConnectCacheDatabase.insertQueryResult(
         authUid.authUid,
-        queryId1.bytes,
+        queryId1.queryId,
         structSample1.struct,
         maxAge = maxAge,
         currentTimeMillis = 0L,
@@ -403,7 +412,7 @@ class DataConnectCacheDatabaseUnitTest {
       )
       dataConnectCacheDatabase.insertQueryResult(
         authUid.authUid,
-        queryId2.bytes,
+        queryId2.queryId,
         structSample2.struct,
         maxAge = maxAge,
         currentTimeMillis = 0L,
@@ -413,7 +422,7 @@ class DataConnectCacheDatabaseUnitTest {
       val result1 =
         dataConnectCacheDatabase.getQueryResult(
           authUid.authUid,
-          queryId1.bytes,
+          queryId1.queryId,
           currentTimeMillis = 0L,
           staleResult = Stale::class,
         )
@@ -421,7 +430,7 @@ class DataConnectCacheDatabaseUnitTest {
       val result2 =
         dataConnectCacheDatabase.getQueryResult(
           authUid.authUid,
-          queryId2.bytes,
+          queryId2.queryId,
           currentTimeMillis = 0L,
           staleResult = Stale::class,
         )
@@ -446,7 +455,7 @@ class DataConnectCacheDatabaseUnitTest {
     ) { authUid, queryId, maxAge, queryResult ->
       dataConnectCacheDatabase.insertQueryResult(
         authUid.authUid,
-        queryId.bytes,
+        queryId.queryId,
         queryResult.hydratedStruct,
         maxAge = maxAge,
         currentTimeMillis = 0L,
@@ -456,7 +465,7 @@ class DataConnectCacheDatabaseUnitTest {
       val result =
         dataConnectCacheDatabase.getQueryResult(
           authUid.authUid,
-          queryId.bytes,
+          queryId.queryId,
           currentTimeMillis = 0L,
           staleResult = Stale::class,
         )
@@ -499,7 +508,7 @@ class DataConnectCacheDatabaseUnitTest {
         queryIds.zip(queryResults).forEach { (queryId, queryResult) ->
           dataConnectCacheDatabase.insertQueryResult(
             authUid.authUid,
-            queryId.bytes,
+            queryId.queryId,
             queryResult.hydratedStruct,
             maxAge = maxAge,
             currentTimeMillis = 0L,
@@ -514,14 +523,13 @@ class DataConnectCacheDatabaseUnitTest {
             val result =
               dataConnectCacheDatabase.getQueryResult(
                 authUid.authUid,
-                queryId.bytes,
+                queryId.queryId,
                 currentTimeMillis = 0L,
                 staleResult = Stale::class,
               )
             val structFromDb = result.shouldBeInstanceOf<Found>().struct
             withClue(
-              "queryIdIndex=$queryIdIndex size=${queryIds.size}, " +
-                "queryId=${queryId.bytes.to0xHexString()}"
+              "queryIdIndex=$queryIdIndex size=${queryIds.size}, " + "queryId=${queryId.queryId}"
             ) {
               structFromDb shouldBe queryResult.hydratedStruct
             }
@@ -550,7 +558,7 @@ class DataConnectCacheDatabaseUnitTest {
       queryIds.zip(queryResults).forEach { (queryId, queryResult) ->
         dataConnectCacheDatabase.insertQueryResult(
           authUid.authUid,
-          queryId.bytes,
+          queryId.queryId,
           queryResult.hydratedStruct,
           maxAge = maxAge,
           currentTimeMillis = 0L,
@@ -562,11 +570,11 @@ class DataConnectCacheDatabaseUnitTest {
         val result =
           dataConnectCacheDatabase.getQueryResult(
             authUid.authUid,
-            queryId.bytes,
+            queryId.queryId,
             currentTimeMillis = 0L,
             staleResult = Stale::class,
           )
-        withClue("index=$index size=${queryIds.size}, queryId=${queryId.bytes.to0xHexString()}") {
+        withClue("index=$index size=${queryIds.size}, queryId=${queryId.queryId}") {
           val structFromDb = result.shouldBeInstanceOf<Found>().struct
           structFromDb shouldBe queryResult.hydratedStruct
         }
@@ -594,7 +602,7 @@ class DataConnectCacheDatabaseUnitTest {
 
       dataConnectCacheDatabase.insertQueryResult(
         authUid.authUid,
-        queryId1.bytes,
+        queryId1.queryId,
         queryResult1.hydratedStruct,
         maxAge = maxAge,
         currentTimeMillis = 0L,
@@ -602,7 +610,7 @@ class DataConnectCacheDatabaseUnitTest {
       )
       dataConnectCacheDatabase.insertQueryResult(
         authUid.authUid,
-        queryId2.bytes,
+        queryId2.queryId,
         queryResult2.hydratedStruct,
         maxAge = maxAge,
         currentTimeMillis = 0L,
@@ -613,7 +621,7 @@ class DataConnectCacheDatabaseUnitTest {
         val result =
           dataConnectCacheDatabase.getQueryResult(
             authUid.authUid,
-            queryId2.bytes,
+            queryId2.queryId,
             currentTimeMillis = 0L,
             staleResult = Stale::class,
           )
@@ -624,7 +632,7 @@ class DataConnectCacheDatabaseUnitTest {
         val result =
           dataConnectCacheDatabase.getQueryResult(
             authUid.authUid,
-            queryId1.bytes,
+            queryId1.queryId,
             currentTimeMillis = 0L,
             staleResult = Stale::class,
           )
@@ -637,18 +645,144 @@ class DataConnectCacheDatabaseUnitTest {
   }
 
   @Test
+  fun `insertQueryResult() should set last_update_sequence_number only for itself`() = runTest {
+    dataConnectCacheDatabase.initialize()
+
+    checkAll(
+      propTestConfig,
+      authUidArb(),
+      Arb.dataConnect.maxAge(min = MIN_NONZERO_DURATION),
+      queryIdArb().distinctPair(),
+    ) { authUid, maxAge, (queryId1, queryId2) ->
+      val queryResultArb =
+        QueryResultArb(
+          entityIdArb = @OptIn(DelicateKotest::class) Arb.proto.structKey(length = 4).distinct(),
+          entityCountRange = 0..5,
+        )
+      val queryResult1 = queryResultArb.bind()
+      val queryResult2 = queryResultArb.bind()
+
+      val sequenceNumber1 =
+        dataConnectCacheDatabase.insertQueryResult(
+          authUid.authUid,
+          queryId1.queryId,
+          queryResult1.hydratedStruct,
+          maxAge = maxAge,
+          currentTimeMillis = 0L,
+          getEntityIdForPath = queryResult1::getEntityIdForPath,
+        )
+      val sequenceNumber2 =
+        dataConnectCacheDatabase.insertQueryResult(
+          authUid.authUid,
+          queryId2.queryId,
+          queryResult2.hydratedStruct,
+          maxAge = maxAge,
+          currentTimeMillis = 0L,
+          getEntityIdForPath = queryResult2::getEntityIdForPath,
+        )
+
+      withClue("query1") {
+        val result =
+          dataConnectCacheDatabase.getQueryResult(
+            authUid.authUid,
+            queryId1.queryId,
+            currentTimeMillis = 0L,
+            staleResult = Stale::class,
+          )
+        check(result is Found)
+        result.maxLastUpdateSequenceNumber shouldBe sequenceNumber1
+      }
+      withClue("query2") {
+        val result =
+          dataConnectCacheDatabase.getQueryResult(
+            authUid.authUid,
+            queryId2.queryId,
+            currentTimeMillis = 0L,
+            staleResult = Stale::class,
+          )
+        check(result is Found)
+        result.maxLastUpdateSequenceNumber shouldBe sequenceNumber2
+      }
+    }
+  }
+
+  @Test
+  fun `insertQueryResult() should update last_update_sequence_number for its entities`() = runTest {
+    dataConnectCacheDatabase.initialize()
+
+    checkAll(
+      propTestConfig,
+      authUidArb(),
+      Arb.dataConnect.maxAge(min = MIN_NONZERO_DURATION),
+      queryIdArb().distinctPair(),
+    ) { authUid, maxAge, (queryId1, queryId2) ->
+      val queryResultArb =
+        QueryResultArb(
+            entityCountRange = 0..5,
+            entityRepeatPolicy = INTER_SAMPLE,
+          )
+          .filter { it.entityByPath.isNotEmpty() }
+      val queryResult1 = queryResultArb.bind()
+      val queryResult2 = queryResultArb.bind()
+
+      dataConnectCacheDatabase.insertQueryResult(
+        authUid.authUid,
+        queryId1.queryId,
+        queryResult1.hydratedStruct,
+        maxAge = maxAge,
+        currentTimeMillis = 0L,
+        getEntityIdForPath = queryResult1::getEntityIdForPath,
+      )
+      val sequenceNumber2 =
+        dataConnectCacheDatabase.insertQueryResult(
+          authUid.authUid,
+          queryId2.queryId,
+          queryResult2.hydratedStruct,
+          maxAge = maxAge,
+          currentTimeMillis = 0L,
+          getEntityIdForPath = queryResult2::getEntityIdForPath,
+        )
+
+      withClue("query1") {
+        val result =
+          dataConnectCacheDatabase.getQueryResult(
+            authUid.authUid,
+            queryId1.queryId,
+            currentTimeMillis = 0L,
+            staleResult = Stale::class,
+          )
+        check(result is Found)
+        // Even query1 should report query2's sequence number since its insert modified the
+        // last_update_sequence_number of at least one of query1's entities.
+        result.maxLastUpdateSequenceNumber shouldBe sequenceNumber2
+      }
+      withClue("query2") {
+        val result =
+          dataConnectCacheDatabase.getQueryResult(
+            authUid.authUid,
+            queryId2.queryId,
+            currentTimeMillis = 0L,
+            staleResult = Stale::class,
+          )
+        check(result is Found)
+        result.maxLastUpdateSequenceNumber shouldBe sequenceNumber2
+      }
+    }
+  }
+
+  @Test
   fun `getQueryResult() when maxAge is zero and staleResult=Stale should return Stale`() =
     testQueryResultInsertedWithMaxAgeZero(
       staleResult = Stale::class,
-      verifyResult = { result, _, expectedStaleness -> result shouldBe Stale(expectedStaleness) }
+      verifyResult = { result, _, expectedStaleness, _ -> result shouldBe Stale(expectedStaleness) }
     )
 
   @Test
   fun `getQueryResult() when maxAge is zero and staleResult=Found should return Found`() =
     testQueryResultInsertedWithMaxAgeZero(
       staleResult = Found::class,
-      verifyResult = { result, insertedData, expectedStaleness ->
-        result shouldBe Found(insertedData, -expectedStaleness)
+      verifyResult = { result, insertedData, expectedStaleness, insertSequenceNumber ->
+        result shouldBe Found(insertedData, -expectedStaleness, insertSequenceNumber)
       }
     )
 
@@ -656,12 +790,18 @@ class DataConnectCacheDatabaseUnitTest {
   fun `getQueryResult() when maxAge is zero and staleResult=NotFound should return NotFound`() =
     testQueryResultInsertedWithMaxAgeZero(
       staleResult = NotFound::class,
-      verifyResult = { result, _, _ -> result shouldBe NotFound }
+      verifyResult = { result, _, _, _ -> result shouldBe NotFound }
     )
 
   private fun testQueryResultInsertedWithMaxAgeZero(
     staleResult: KClass<out GetQueryResultResult>,
-    verifyResult: (GetQueryResultResult, insertedData: Struct, expectedStaleness: Duration) -> Unit,
+    verifyResult:
+      (
+        GetQueryResultResult,
+        insertedData: Struct,
+        expectedStaleness: Duration,
+        sequenceNumber: SqliteSequenceNumber,
+      ) -> Unit,
   ) = runTest {
     dataConnectCacheDatabase.initialize()
 
@@ -681,9 +821,9 @@ class DataConnectCacheDatabaseUnitTest {
         time1 = time1,
         time2 = time2,
         staleResult = staleResult,
-        verifyResult = {
+        verifyResult = { result, insertSequenceNumber ->
           val expectedStaleness = durationFromMillis(time2.toBigInteger() - time1.toBigInteger())
-          verifyResult(it, queryResult.hydratedStruct, expectedStaleness)
+          verifyResult(result, queryResult.hydratedStruct, expectedStaleness, insertSequenceNumber)
         },
       )
     }
@@ -693,15 +833,15 @@ class DataConnectCacheDatabaseUnitTest {
   fun `getQueryResult() after maxAge time has passed and staleResult=Stale should return Stale`() =
     testGetQueryResultAfterMaxAgeTimeHasPassed(
       staleResult = Stale::class,
-      verifyResult = { result, _, expectedStaleness -> result shouldBe Stale(expectedStaleness) }
+      verifyResult = { result, _, expectedStaleness, _ -> result shouldBe Stale(expectedStaleness) }
     )
 
   @Test
   fun `getQueryResult() after maxAge time has passed and staleResult=Found should return Found`() =
     testGetQueryResultAfterMaxAgeTimeHasPassed(
       staleResult = Found::class,
-      verifyResult = { result, insertedData, expectedStaleness ->
-        result shouldBe Found(insertedData, -expectedStaleness)
+      verifyResult = { result, insertedData, expectedStaleness, insertSequenceNumber ->
+        result shouldBe Found(insertedData, -expectedStaleness, insertSequenceNumber)
       }
     )
 
@@ -709,12 +849,18 @@ class DataConnectCacheDatabaseUnitTest {
   fun `getQueryResult() after maxAge time has passed and staleResult=NotFound should return NotFound`() =
     testGetQueryResultAfterMaxAgeTimeHasPassed(
       staleResult = NotFound::class,
-      verifyResult = { result, _, _ -> result shouldBe NotFound }
+      verifyResult = { result, _, _, _ -> result shouldBe NotFound }
     )
 
   private fun testGetQueryResultAfterMaxAgeTimeHasPassed(
     staleResult: KClass<out GetQueryResultResult>,
-    verifyResult: (GetQueryResultResult, insertedData: Struct, expectedStaleness: Duration) -> Unit,
+    verifyResult:
+      (
+        GetQueryResultResult,
+        insertedData: Struct,
+        expectedStaleness: Duration,
+        insertSequenceNumber: SqliteSequenceNumber,
+      ) -> Unit,
   ) = runTest {
     dataConnectCacheDatabase.initialize()
 
@@ -757,7 +903,14 @@ class DataConnectCacheDatabaseUnitTest {
         time1 = time1,
         time2 = time2,
         staleResult = staleResult,
-        verifyResult = { verifyResult(it, queryResult.hydratedStruct, expectedStaleness) },
+        verifyResult = { result, insertSequenceNumber ->
+          verifyResult(
+            result,
+            queryResult.hydratedStruct,
+            expectedStaleness,
+            insertSequenceNumber,
+          )
+        },
       )
     }
   }
@@ -792,7 +945,14 @@ class DataConnectCacheDatabaseUnitTest {
         time1 = time1,
         time2 = time2,
         staleResult = staleResult,
-        verifyResult = { it shouldBe Found(queryResult.hydratedStruct, Duration.ZERO) },
+        verifyResult = { result, insertSequenceNumber ->
+          result shouldBe
+            Found(
+              queryResult.hydratedStruct,
+              Duration.ZERO,
+              insertSequenceNumber,
+            )
+        },
       )
     }
   }
@@ -843,7 +1003,14 @@ class DataConnectCacheDatabaseUnitTest {
           time1 = time1,
           time2 = time2,
           staleResult = staleResult,
-          verifyResult = { it shouldBe Found(queryResult.hydratedStruct, freshnessRemaining) },
+          verifyResult = { result, insertSequenceNumber ->
+            result shouldBe
+              Found(
+                queryResult.hydratedStruct,
+                freshnessRemaining,
+                insertSequenceNumber,
+              )
+          },
         )
       }
     }
@@ -856,26 +1023,194 @@ class DataConnectCacheDatabaseUnitTest {
     time1: Long,
     time2: Long,
     staleResult: KClass<out GetQueryResultResult>,
-    verifyResult: (GetQueryResultResult) -> Unit,
+    verifyResult: (GetQueryResultResult, insertSequenceNumber: SqliteSequenceNumber) -> Unit,
   ) {
-    dataConnectCacheDatabase.insertQueryResult(
-      authUid.authUid,
-      queryId.bytes,
-      queryResultData,
-      maxAge = maxAge,
-      currentTimeMillis = time1,
-      getEntityIdForPath = null,
-    )
+    val sequenceNumber =
+      dataConnectCacheDatabase.insertQueryResult(
+        authUid.authUid,
+        queryId.queryId,
+        queryResultData,
+        maxAge = maxAge,
+        currentTimeMillis = time1,
+        getEntityIdForPath = null,
+      )
 
     val result =
       dataConnectCacheDatabase.getQueryResult(
         authUid.authUid,
-        queryId.bytes,
+        queryId.queryId,
         currentTimeMillis = time2,
         staleResult,
       )
 
-    verifyResult(result)
+    verifyResult(result, sequenceNumber)
+  }
+
+  @Test
+  fun `getQueryResult() last_update_sequence_number all null`() =
+    testGetQueryResultReturnsCorrectMaxLastUpdateSequenceNumber(
+      editDbAfterInsert = { db, _, logger ->
+        db.execSQL(logger, "UPDATE queries SET last_update_sequence_number = NULL")
+        db.execSQL(logger, "UPDATE entities SET last_update_sequence_number = NULL")
+      },
+      verifyResult = { result, _, _ -> result.maxLastUpdateSequenceNumber.shouldBeNull() },
+    )
+
+  @Test
+  fun `getQueryResult() query last_update_sequence_number is null`() =
+    testGetQueryResultReturnsCorrectMaxLastUpdateSequenceNumber(
+      editDbAfterInsert = { db, _, logger ->
+        db.execSQL(logger, "UPDATE queries SET last_update_sequence_number = NULL")
+      },
+      verifyResult = { result, insertSequenceNumber, _ ->
+        result.maxLastUpdateSequenceNumber shouldBe insertSequenceNumber
+      },
+    )
+
+  @Test
+  fun `getQueryResult() entities last_update_sequence_number are null`() =
+    testGetQueryResultReturnsCorrectMaxLastUpdateSequenceNumber(
+      editDbAfterInsert = { db, _, logger ->
+        db.execSQL(logger, "UPDATE entities SET last_update_sequence_number = NULL")
+      },
+      verifyResult = { result, insertSequenceNumber, _ ->
+        result.maxLastUpdateSequenceNumber shouldBe insertSequenceNumber
+      },
+    )
+
+  @Test
+  fun `getQueryResult() query last_update_sequence_number is largest`() =
+    testGetQueryResultReturnsCorrectMaxLastUpdateSequenceNumber(
+      editDbAfterInsert = { db, insertSequenceNumber, logger ->
+        val newSequenceNumber = insertSequenceNumber.sequenceNumber + 1
+        db.execSQL(logger, "UPDATE queries SET last_update_sequence_number = $newSequenceNumber")
+        SqliteSequenceNumber(newSequenceNumber)
+      },
+      verifyResult = { result, _, querySequenceNumber ->
+        result.maxLastUpdateSequenceNumber shouldBe querySequenceNumber
+      },
+    )
+
+  @Test
+  fun `getQueryResult() entity last_update_sequence_number is largest`() =
+    testGetQueryResultReturnsCorrectMaxLastUpdateSequenceNumber(
+      editDbAfterInsert = { db, _, logger ->
+        val entityRowIds =
+          db.rawQuery(logger, "SELECT id FROM entities") { cursor ->
+            buildList {
+              while (cursor.moveToNext()) {
+                add(cursor.getLong(0))
+              }
+            }
+          }
+
+        val sequenceNumberArb = Arb.longWithEvenNumDigitsDistribution()
+        val sequenceNumbers = List(entityRowIds.size + 1) { sequenceNumberArb.bind() }
+        val distinctSortedSequenceNumbers = sequenceNumbers.distinct().sorted()
+        assume(distinctSortedSequenceNumbers.size >= 2)
+
+        val querySequenceNumber =
+          distinctSortedSequenceNumbers.dropLast(1).random(randomSource().random)
+        db.execSQL(logger, "UPDATE queries SET last_update_sequence_number = $querySequenceNumber")
+
+        val entitySequenceNumbers = sequenceNumbers.minus(querySequenceNumber).iterator()
+        entityRowIds.forEach { entityRowId ->
+          val entitySequenceNumber = entitySequenceNumbers.next()
+          db.execSQL(
+            logger,
+            """
+          UPDATE entities
+          SET last_update_sequence_number = $entitySequenceNumber
+          WHERE id=$entityRowId
+        """
+          )
+        }
+
+        SqliteSequenceNumber(distinctSortedSequenceNumbers.last())
+      },
+      verifyResult = { result, _, maxEntitySequenceNumber ->
+        result.maxLastUpdateSequenceNumber shouldBe maxEntitySequenceNumber
+      },
+    )
+
+  private fun <T> testGetQueryResultReturnsCorrectMaxLastUpdateSequenceNumber(
+    editDbAfterInsert: PropertyContext.(SQLiteDatabase, SqliteSequenceNumber, Logger) -> T,
+    verifyResult: (Found, SqliteSequenceNumber, T) -> Unit,
+  ) = runTest {
+    dataConnectCacheDatabase.initialize()
+
+    checkAll(
+      propTestConfig,
+      authUidArb(),
+      queryIdArb(),
+      Arb.dataConnect.maxAge(min = MIN_NONZERO_DURATION),
+      QueryResultArb(entityCountRange = 1..5),
+    ) { authUid, queryId, maxAge, queryResult ->
+      val dbFile = File(temporaryFolder.newFolder(), "db.sqlite")
+      val mockLogger: Logger = mockk(relaxed = true)
+
+      // Insert a query and entities into the database
+      val insertSequenceNumber = run {
+        val dataConnectCacheDatabase =
+          DataConnectCacheDatabase(
+            dbFile,
+            Dispatchers.Default,
+            mockLogger,
+          )
+        try {
+          dataConnectCacheDatabase.initialize()
+          dataConnectCacheDatabase.insertQueryResult(
+            authUid.authUid,
+            queryId.queryId,
+            queryResult.hydratedStruct,
+            maxAge = maxAge,
+            currentTimeMillis = 0L,
+            getEntityIdForPath = queryResult::getEntityIdForPath,
+          )
+        } finally {
+          dataConnectCacheDatabase.close()
+        }
+      }
+
+      // Manually edit the "last_update_sequence_number" column in the database
+      val editDbAfterInsertResult = run {
+        DataConnectSQLiteDatabaseOpener.open(dbFile, mockLogger).use { db ->
+          db.beginTransaction()
+          try {
+            val result = editDbAfterInsert(db, insertSequenceNumber, mockLogger)
+            db.setTransactionSuccessful()
+            result
+          } finally {
+            db.endTransaction()
+          }
+        }
+      }
+
+      // Load the query from the database
+      val result = run {
+        val dataConnectCacheDatabase =
+          DataConnectCacheDatabase(
+            dbFile,
+            Dispatchers.Default,
+            mockLogger,
+          )
+        try {
+          dataConnectCacheDatabase.initialize()
+          dataConnectCacheDatabase.getQueryResult(
+            authUid.authUid,
+            queryId.queryId,
+            currentTimeMillis = 0L,
+            staleResult = Stale::class,
+          )
+        } finally {
+          dataConnectCacheDatabase.close()
+        }
+      }
+
+      // Verify that the maxLastUpdateSequenceNumber was the expected value
+      check(result is Found)
+      verifyResult(result, insertSequenceNumber, editDbAfterInsertResult)
+    }
   }
 }
 
@@ -894,12 +1229,14 @@ private fun authUidArb(): Arb<AuthUidSample> =
     AuthUidSample(if (it !== null) AuthUid(it) else null)
   }
 
-private data class QueryIdSample(val bytes: ImmutableByteArray) {
-  override fun toString() = "QueryIdSample(bytes=${bytes.to0xHexString()})"
+private data class QueryIdSample(val queryId: QueryId) {
+  override fun toString() = "QueryIdSample(queryId=$queryId)"
 }
 
 private fun queryIdArb(): Arb<QueryIdSample> =
-  Arb.byteArray(Arb.int(1..25), Arb.byte()).map { QueryIdSample(ImmutableByteArray.adopt(it)) }
+  Arb.byteArray(Arb.int(1..25), Arb.byte()).map {
+    QueryIdSample(QueryId(ImmutableByteArray.adopt(it)))
+  }
 
 internal fun QueryResultArb.Sample.hydratedStructWithMutatedEntityValuesFrom(
   other: QueryResultArb.Sample