+++ firebase-dataconnect/CHANGELOG.md
@@ -9,6 +9,12 @@
   as the random numbers were not used in a security-sensitive context,
   thus the performance costs of secure random number generation were unnecessary.
   ([#8154](https://github.com/firebase/firebase-android-sdk/pull/8154))
+- [fixed] Queries executed with FetchPolicy.CACHE_ONLY now fail, as expected,
+  if local caching is not enabled, instead of behaving like SERVER_ONLY.
+  ([#8214](https://github.com/firebase/firebase-android-sdk/pull/8214))
+- [changed] Realtime query results now update the local cache as query
+  results are received.
+  ([#8220](https://github.com/firebase/firebase-android-sdk/pull/8220))
 
 # 17.2.2